# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActionItemPriority =
  (typeof ActionItemPriority)[keyof typeof ActionItemPriority];

export const ActionItemPriority = {
  High: "High",
  Medium: "Medium",
  Low: "Low",
} as const;

export interface ActionItem {
  owner: string;
  task: string;
  dueDate: string;
  priority: ActionItemPriority;
}

export interface TranscriptSegment {
  start: number;
  end: number;
  text: string;
}

export interface Meeting {
  id: number;
  meetingName: string;
  fileId: string;
  transcript: string;
  segments?: TranscriptSegment[];
  durationSec: number;
  costUsd: number;
  createdAt: string;
  summary?: string | null;
  actionItems?: ActionItem[] | null;
  decisions?: string[] | null;
  openQuestions?: string[] | null;
  analysisCostUsd?: number | null;
}

export interface MeetingStats {
  totalMeetings: number;
  totalDurationSec: number;
  totalCostUsd: number;
  avgDurationSec: number;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UploadAudioBody = {
  audio_blob: Blob;
  meeting_name: string;
};