# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: meetings
    description: Meeting recording and transcription operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List all meetings
      description: Returns all saved meeting transcriptions
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a meeting by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/upload:
    post:
      operationId: uploadAudio
      tags: [meetings]
      summary: Upload audio for transcription
      description: Accepts a WebM audio blob and meeting name, converts to WAV, transcribes via Whisper, saves meeting record
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio_blob:
                  type: string
                  format: binary
                meeting_name:
                  type: string
              required:
                - audio_blob
                - meeting_name
      responses:
        "200":
          description: Transcription complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/{id}/download:
    get:
      operationId: downloadTranscript
      tags: [meetings]
      summary: Download transcript as markdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Markdown file
          content:
            text/markdown:
              schema:
                type: string
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/{id}/analyze:
    post:
      operationId: analyzeMeeting
      tags: [meetings]
      summary: Run AI analysis on a meeting transcript
      description: Uses GPT-4o to generate summary, action items, decisions, and open questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis complete — returns full meeting with analysis fields populated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/{id}/export:
    get:
      operationId: exportMeeting
      tags: [meetings]
      summary: Export meeting as Excel workbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/stats:
    get:
      operationId: getMeetingStats
      tags: [meetings]
      summary: Get aggregated meeting statistics
      description: Total meetings, total duration, total cost, average duration
      responses:
        "200":
          description: Meeting stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ActionItem:
      type: object
      properties:
        owner:
          type: string
        task:
          type: string
        dueDate:
          type: string
        priority:
          type: string
          enum: [High, Medium, Low]
      required:
        - owner
        - task
        - dueDate
        - priority

    Meeting:
      type: object
      properties:
        id:
          type: integer
        meetingName:
          type: string
        fileId:
          type: string
        transcript:
          type: string
        segments:
          type: array
          items:
            $ref: "#/components/schemas/TranscriptSegment"
        durationSec:
          type: number
        costUsd:
          type: number
        createdAt:
          type: string
          format: date-time
        summary:
          type: string
          nullable: true
        actionItems:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ActionItem"
        decisions:
          type: array
          nullable: true
          items:
            type: string
        openQuestions:
          type: array
          nullable: true
          items:
            type: string
        analysisCostUsd:
          type: number
          nullable: true
      required:
        - id
        - meetingName
        - fileId
        - transcript
        - durationSec
        - costUsd
        - createdAt

    TranscriptSegment:
      type: object
      properties:
        start:
          type: number
        end:
          type: number
        text:
          type: string
      required:
        - start
        - end
        - text

    MeetingStats:
      type: object
      properties:
        totalMeetings:
          type: integer
        totalDurationSec:
          type: number
        totalCostUsd:
          type: number
        avgDurationSec:
          type: number
      required:
        - totalMeetings
        - totalDurationSec
        - totalCostUsd
        - avgDurationSec

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success