# lib/api-zod/src/generated/types/actionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionItemPriority } from "./actionItemPriority";

export interface ActionItem {
  owner: string;
  task: string;
  dueDate: string;
  priority: ActionItemPriority;
}