# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionItem } from "./actionItem";
import type { TranscriptSegment } from "./transcriptSegment";

export interface Meeting {
  id: number;
  meetingName: string;
  fileId: string;
  transcript: string;
  segments?: TranscriptSegment[];
  durationSec: number;
  costUsd: number;
  createdAt: Date;
  summary?: string | null;
  actionItems?: ActionItem[] | null;
  decisions?: string[] | null;
  openQuestions?: string[] | null;
  analysisCostUsd?: number | null;
}